+++ NEWS.md
@@ -1,6 +1,8 @@
 # hellmer 0.1.2
 
-## Breaking changes
+## Lifecycle changes
+* Removed the timeout feature as it's better handled by `option(ellmer_timeout_s = 120)` in ellmer 0.1.1
+
 * Moved parameters from `chat_sequential` and `chat_future` to `$batch()`, except for `chat_model` and `...`.
 
 * `$batch()` gains `progress` in addition to  `echo` and `...` which are passed to the chat call.